// src/issues/interfaces/label.ts
// Generated by https://quicktype.io

export interface Label {
    id:          number;
    node_id:     string;
    url:         string;
    name:        string;
    color:       string;
    default:     boolean;
    description?: string;
}

export enum Description {
    Empty = "",
    PullRequestsThatUpdateADependencyFile = "Pull requests that update a dependency file",
}